refactor: renomeia indexes para melhorar a identificação
