+++ eco-revive/src/components/copyright/copyright.js
