+++ eco-revive/src/pages/guidePage/components/ProcedureBox/procedureBox.js
