+++ eco-revive/src/pages/guidePage/guidepage.js
@@ -4,9 +4,9 @@ import BottomWhiteWave from "../../components/bottomWhiteWave";
 import BottomBlueWave from "../../components/bottomBlueWave";
 import TopBlueWave from "../../components/TopBlueWave";
 import TopGreenWave from "../../components/TopGreenWave";
-import ProcedureBox from "./components/ProcedureBox";
-import ColaboratorSection from "../../components/colaboratorSection";
-import Copyright from "../../components/copyright";
+import ProcedureBox from "./components/ProcedureBox/procedureBox";
+import ColaboratorSection from "../../components/colaboratorSection/colaboratorSection";
+import Copyright from "../../components/copyright/copyright";
 
 import Logo from "../../assets/Logo320.png"; //rever qualidade da imagem
 import RecycleSymbol from "../../assets/Recycle.png";
